+++ frontend/src/editor/EditorMenu.svelte
@@ -1,7 +1,8 @@
 <script>
   import { createEventDispatcher } from "svelte";
 
-  import { _, urlFor } from "../helpers";
+  import { _ } from "../i18n";
+  import { urlFor } from "../helpers";
 
   /** @type {string[]} */
   export let sources;
