+++ frontend/src/entry-forms/AccountInput.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { _ } from "../helpers";
+  import { _ } from "../i18n";
   import { accounts } from "../stores";
 
   import AutocompleteInput from "../AutocompleteInput.svelte";
