+++ frontend/src/entry-forms/Balance.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { _ } from "../helpers";
+  import { _ } from "../i18n";
   import { currencies } from "../stores";
 
   import AutocompleteInput from "../AutocompleteInput.svelte";
