+++ frontend/src/stores/chart.ts
@@ -1,6 +1,6 @@
 import { writable, derived } from "svelte/store";
 
-import { _ } from "../helpers";
+import { _ } from "../i18n";
 import iso4217currencies from "../lib/iso4217";
 import { currencies_sorted, operating_currency } from ".";
 import { NamedChartTypes } from "../charts";
